requirements.txt
pytest==9.0.2
playwright==1.39.1
requests==2.32.0
Pillow==10.0.0
pytest-xdist==3.2.2